@@ -8,26 +8,26 @@
     "tauri": "tauri"
   },
   "dependencies": {
-    "@cryptopuppie/useeoschain": "^0.1.2",
-    "@fontsource/karla": "^4.5.0",
-    "@headlessui/react": "^1.4.2",
-    "@tauri-apps/api": "^1.0.0-beta.8",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-syntax-highlighter": "^15.4.5",
-    "url-join": "^4.0.1"
+    "@cryptopuppie/useeoschain": "^0.2.0",
+    "@fontsource/karla": "^4.5.9",
+    "@headlessui/react": "^1.6.6",
+    "@tauri-apps/api": "^1.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-syntax-highlighter": "^15.5.0",
+    "url-join": "^5.0.0"
   },
   "devDependencies": {
-    "@tauri-apps/cli": "^1.0.0-beta.10",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@types/react-syntax-highlighter": "^13.5.2",
+    "@tauri-apps/cli": "^1.0.5",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-syntax-highlighter": "^15.5.4",
     "@types/url-join": "^4.0.1",
-    "@vitejs/plugin-react": "^1.1.4",
-    "autoprefixer": "^10.4.2",
-    "postcss": "^8.4.5",
-    "tailwindcss": "^3.0.13",
-    "typescript": "^4.5.4",
-    "vite": "^2.7.11"
+    "@vitejs/plugin-react": "^2.0.1",
+    "autoprefixer": "^10.4.8",
+    "postcss": "^8.4.16",
+    "tailwindcss": "^3.1.8",
+    "typescript": "^4.7.4",
+    "vite": "^3.0.8"
   }
 }
