@@ -1,7 +1,7 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
-import './index.css'
+import ReactDOM from 'react-dom/client'
 import App from './App'
+import './index.css'
 
 import "@fontsource/karla/200.css"
 import "@fontsource/karla/300.css"
@@ -13,9 +13,8 @@ import "@fontsource/karla/800.css"
 // import "@fontsource/karla/900.css"
 
 
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 )
